requirements.txt
python-telegram-bot==12.2.0
requests==2.21.0
urllib3==1.24.3